Jekyll (#3)
* move site files to root directory

The default configuration of jekyll expects this.

* add Gemfile for jekyll

* jekyllfy the site

So far the content of the site has not been changed. In the near future
I am sure there will be a blog post and at that time it will be easy to
add. There will need to be a few changes make to the layout and homepage
to accomidate a blog post.

* add jekyll directions to README